+++ resources/views/components/input/partials/ckeditor-base.blade.php
@@ -2,18 +2,6 @@
 
 {{ $slot }}
 
-@assets('scripts')
-    <script src="{{ asset('js/ckeditor.js') }}"></script>
-
-    <style>
-        .ck.ck-balloon-panel.ck-balloon-panel_position_border-side_right.ck-powered-by-balloon {
-            display: none;
-        }
-
-        .ck.ck-editor__editable_inline {
-            background: var(--ck-color-base-background);
-            border: 1px solid var(--ck-color-base-border);
-        }
-    </style>
-@endassets
-
+@pushOnce('scripts')
+    <script src="{{ asset('js/nk-ckeditor.js') }}"></script>
+@endPushOnce